chore: move error_on_miss into advanced fields
